Use SPI for DriverStateContext
