+++ shardingsphere-jdbc/shardingsphere-jdbc-core/src/main/resources/META-INF/services/org.apache.shardingsphere.driver.state.DriverState
@@ -15,6 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.driver.governance.internal.state.impl.OKDriverState
-org.apache.shardingsphere.driver.governance.internal.state.impl.CircuitBreakDriverState
-org.apache.shardingsphere.driver.governance.internal.state.impl.LockDriverState
+org.apache.shardingsphere.driver.state.OKDriverState
